requirements.txt
Flask==3.0.0
flask-cors==6.0.0
Flask_JWT_Extended==4.6.0
Flask_Limiter==3.9.2
flask_swagger_ui==4.11.1
pandas==2.2.3
pydantic==1.10.18
pydantic_core==2.23.3
pydantic-env-settings
pyodbc==5.1.0
Requests==2.32.3
cmd_gui_kit==1.0.0
termcolor==2.3.0
python-dotenv==1.0.1
email_validator==2.2.0
plotly==5.24.1
bcrypt==4.1.3
Werkzeug==3.1.3
isodate==0.7.2
flask-talisman==1.1.0
psycopg2-binary==2.9.9

+# relational drivers
 psycopg2-binary
 pymysql

+# NoSQL drivers
 pymongo
 google-cloud-firestore